chore: remove repository interfaces from Domain project
