+++ Commands/Handlers/BankAccount/AddBankAccount/AddBankAccountHandler.cs
@@ -1,7 +1,7 @@
 using Commands.Extensions;
 using Commands.Services;
 
-using Domain.Interfaces;
+using Persistence.Repositories;
 
 namespace Commands.Handlers.BankAccount.AddBankAccount;
