+++ Commands/Handlers/Member/DeleteMember/DeleteMemberCommand.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Interfaces;
+﻿using Persistence.Repositories;
 
 namespace Commands.Handlers.Member.DeleteMember;
