+++ Commands/Handlers/Member/ExtendMembership/ExtendMembershipHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Commands.Extensions;
 using Commands.Services;
 
-using Domain.Interfaces;
+using Persistence.Repositories;
 
 namespace Commands.Handlers.Member.ExtendMembership;
