--- Domain/Interfaces/IMemberRepository.cs
@@ -1,11 +1,0 @@
-namespace Domain.Interfaces;
-
-public interface IMemberRepository
-{
-    Task<Member?> GetById(Guid id);
-    Task<Member?> GetByEmail(string email);
-    Task<IEnumerable<Member>> GetAllAsync();
-    void Add(Member member);
-    void Remove(Member member);
-    Task Save(CancellationToken cancellationToken);
-}
