+++ Persistence/Repositories/BankAccountRepository.cs
@@ -1,5 +1,4 @@
 using Domain;
-using Domain.Interfaces;
 
 using Microsoft.EntityFrameworkCore;
 
@@ -40,4 +39,13 @@ public class BankAccountRepository : IBankAccountRepository
 
         await _context.SaveChangesAsync(ct);
     }
+}
+
+public interface IBankAccountRepository
+{
+    Task<BankAccount?> GetByIdAsync(Guid id, CancellationToken ct);
+    Task<IEnumerable<BankAccount>> GetAllAsync(CancellationToken ct);
+    void Add(BankAccount account);
+    void Remove(BankAccount account);
+    Task SaveAsync(CancellationToken ct);
 }
