+++ Persistence/Repositories/MemberRepository.cs
@@ -1,5 +1,4 @@
 using Domain;
-using Domain.Interfaces;
 
 using Microsoft.EntityFrameworkCore;
 
@@ -44,4 +43,14 @@ public class MemberRepository : IMemberRepository
         await _context.SaveChangesAsync(cancellationToken);
     }
 
+}
+
+public interface IMemberRepository
+{
+    Task<Member?> GetById(Guid id);
+    Task<Member?> GetByEmail(string email);
+    Task<IEnumerable<Member>> GetAllAsync();
+    void Add(Member member);
+    void Remove(Member member);
+    Task Save(CancellationToken cancellationToken);
 }
