+++ Web/Pages/Transactions/TransactionForm.razor
@@ -4,7 +4,7 @@
 @using Queries.Members.Handlers.GetMemberById
 @using Queries.Members.ViewModels
 @using Queries.Members.Handlers.AutocompleteMember
-@using Commands.Handlers.Transaction.GetAttachment
+@using Queries.Transactions.GetAttachment
 @inject IWebHostEnvironment Environment
 
 @inject IDialogService dialogService
